+++ src/main/java/chess/controller/console/dto/position/MovablePathResponseDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto.board;
+package chess.controller.console.dto.position;
 
 import chess.domain.board.position.Path;
