+++ src/main/java/chess/controller/web/dto/move/MovablePathRequestDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto.board;
+package chess.controller.web.dto.move;
 
 public class MovablePathRequestDto {
