+++ src/main/java/chess/controller/web/dto/move/MoveRequestDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto.board;
+package chess.controller.web.dto.move;
 
 public class MoveRequestDto {
